+++ src/common/src/net/abstractfetcher.h
@@ -13,8 +13,8 @@ class AbstractFetcher : public QObject
     Q_PROPERTY(QString source READ source WRITE setSource NOTIFY sourceChanged)
     Q_PROPERTY(NextcloudSettingsBase* settings READ settings WRITE setSettings NOTIFY settingsChanged)
     Q_PROPERTY(CommandQueue* commandQueue READ commandQueue WRITE setCommandQueue NOTIFY commandQueueChanged)
-    Q_PROPERTY(qint16 width READ width WRITE setWidth NOTIFY widthChanged)
-    Q_PROPERTY(qint16 height READ height WRITE setHeight NOTIFY heightChanged)
+    Q_PROPERTY(int width READ width WRITE setWidth NOTIFY widthChanged)
+    Q_PROPERTY(int height READ height WRITE setHeight NOTIFY heightChanged)
     Q_PROPERTY(bool fetching READ fetching NOTIFY fetchingChanged)
 
 public:
@@ -30,17 +30,17 @@ protected:
     void setCommandQueue(CommandQueue* v);
     QString source();
     void setSource(const QString& v);
-    qint16 width();
-    void setWidth(qint16 v);
-    qint16 height();
-    void setHeight(qint16 v);
+    int width();
+    void setWidth(int v);
+    int height();
+    void setHeight(int v);
     bool fetching();
     void setFetching(bool v);
 
 private:
     QString m_source;
-    qint16 m_width = 0;
-    qint16 m_height = 0;
+    int m_width = 0;
+    int m_height = 0;
     bool m_fetching = false;
     NextcloudSettingsBase* m_settings = Q_NULLPTR;
     CommandQueue* m_commandQueue = Q_NULLPTR;
